chore: move conf/conf.go to internal/conf/conf.go
